pcells: Restore transfer view compatibility with dCache 2.12
dCache 2.12 uses transfer IDs derived from the current time. Thus
an integer is not enough to represent these IDs. pcells erroneously
interprets the serial ID as an integer, making it incompatible with
dCache 2.12.

The patch also fixes a parsing issue caused by status messages containing
spaces. The patch uses the Args utility class from dCache to parse the
transfer line, thus ensuring that it can deal with quoted values. Note
that an update to dCache is required for it to provide such quoted
values.

The change is backwards compatible in that any reply from older
versions of dCache is treated as before.

Acked-by: Paul Millar <paul.millar@desy.de>
Patch: https://rb.dcache.org/r/8017/