pcells: Reduce dependency to dcache-vehicles
pcells doesn't need the full dCache dependency. Rather than pulling in 50 MB we now only pull
in 15 MB of jars.

Requires a few patches to dCache before this can be committed.

The patch also removes a few tests for TimeoutCacheException. The classic way to report a
timeout is to return a null message. Some versions of dCache didn't adhear to this, leading
to the use of TimeoutCacheException. Recent versions of dCache once again return null as
expected by the majority of pcells.

Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8083/